Add files via upload
this C++ program implements a number guessing game. It generates a random number using the rand() function from the <cstdlib> library and prompts the player to guess the number. It provides feedback on whether the guess is too high or too low. The game continues until the player guesses the correct number. Libraries used: <iostream> for input/output, <cstdlib> for random number generation, and <ctime> for seeding the random number generator.